backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
mysql-connector-python==8.4.0
pydantic==2.7.1